backend/requirements.txt
Flask==3.1.0
Flask-CORS==6.0.1
openai==0.28.0
python-dotenv==1.0.1